Make key argument const in hashmap_get and hashmap_remove[_free]
This allows const types to be passed to hashmap_get, hashmap_remove and
hashmap_remove_free without a cast or a compiler warning (or an error
in C++).

Originally I wanted to make hashmap keys const altogether (so you could
also hashmap_set with a const void* for example), but then I realized
that would go against the usecase of freeing a key from within a
hashmap_callback. So this commit only affects functions that don't
store the key for the user to use later.