Change callback key to const to match rest of API
The only reason why it was not const that I can see is to so
it can be freed in a callback. The irony is that value is passed
as a `uintptr_t` which is a constant value but still might need
to be freed.

If the key does need to be freed in a callback, it can be cast
to `void *` just like the value.

This is a better fix to #8.